Add files via upload
Welcome to this "Python Magic: A Beginner's Guide to Python's Charm' notebook, your essential guide to mastering Python for data analysis and machine learning. Whether you're a beginner or looking to refine your Python skills, this comprehensive notebook is designed to empower you with the knowledge and tools needed to excel in the world of data science.